Add CeraQuery class
The Cera search routine returns an instance of the CeraQuery class. The representation of this instance is the number of unique entries per variable in the search result.
The class provides a `df` attribute, which represents the search result in form of a pd.DataFrame, with the columns being the variables and the according assets (usually the path to a netCDF file) from the search result.